Move results context to proper folder
