+++ src/pages/Results/shared/context.ts
@@ -1,4 +1,4 @@
 import {createContext} from "react";
-import {ClassModel} from "../../../../shared/EntityTypes.ts";
+import {ClassModel} from "../../../shared/EntityTypes.ts";
 
 export const SelectedClassContext = createContext<ClassModel|null>(null)
